A pair of functions that cache the inverse of a matrix
This first code helps in creating matrix which then sets it into list.
The list contains the functions to set the matrix, get the matrix, set
the inverse and get the inverse.
This second code checks if the matrix inverse is present already. If not
then it inverses, the matrix and shows the required answer else it
retrieves the already calculated inverse.